requirements.txt
dramatiq[redis]     
docling             
python-dotenv      
pymupdf            
watchdog           
torch      
torchvision
onnxruntime